Add argparse module and output filename
Using the argparse module adds help options and makes it easier to
specify options. The `-o` or `--output` option was added to specify the
output filename. This is especially useful when calling `graphdeps.py`
from another Python script.

The default name of `deps.png` remains, so there is no noticeable change
to users who continue to use the script the same way.